backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
bcrypt==4.1.2
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
ecdsa==0.19.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.1.3
fastapi==0.104.1
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
httptools==0.6.1
idna==3.4
Jinja2==3.1.3
joblib==1.4.0
kaggle==1.6.8
keras==3.2.1
langcodes==3.3.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
murmurhash==1.0.10
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pgvector==0.2.5
pillow==10.3.0
preshed==3.0.9
protobuf==4.25.3
psycopg==3.1.18
pyasn1==0.6.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.27.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
